+++ dtslint/ts3.5/TaskDecoder.ts
@@ -1,37 +1,46 @@
 import * as _ from '../../src/TaskDecoder'
-import * as DE from '../../src/DecodeError'
-import * as FS from '../../src/FreeSemigroup'
-import { pipe } from 'fp-ts/lib/pipeable'
 
-// $ExpectType TaskDecoder<{ a: string; b: { c: number; }; }>
-const A = _.type({
-  a: _.string,
-  b: _.type({
-    c: _.number
-  })
+declare const string: _.TaskDecoder<string, string>
+declare const NumberFromString: _.TaskDecoder<string, number>
+
+//
+// type
+//
+
+// $ExpectType TaskDecoder<{ a: string; }, { a: number; }>
+_.type({
+  a: NumberFromString
 })
 
-// $ExpectType TaskDecoder<Partial<{ a: string; b: Partial<{ c: number; }>; }>>
+//
+// partial
+//
+
+// $ExpectType TaskDecoder<{ a: string; }, Partial<{ a: number; }>>
 _.partial({
-  a: _.string,
-  b: _.partial({
-    c: _.number
-  })
+  a: NumberFromString
 })
 
 //
-// TypeOf
+// tuple
 //
 
-// $ExpectType { a: string; b: { c: number; }; }
-export type A = _.TypeOf<typeof A>
+// $ExpectType TaskDecoder<[string], [number]>
+_.tuple(NumberFromString)
 
 //
-// mapLeftWithInput
+// sum
 //
 
-// $ExpectType TaskDecoder<number>
-pipe(
-  _.number,
-  _.mapLeftWithInput((u) => FS.of(DE.leaf(u, 'not a number')))
-)
+// $ExpectType TaskDecoder<{ _tag: unknown; a: string; } | { _tag: unknown; b: string; }, { _tag: "A"; a: number; } | { _tag: "B"; b: number; }>
+_.sum('_tag')({
+  A: _.type({ _tag: _.literal('A'), a: NumberFromString }),
+  B: _.type({ _tag: _.literal('B'), b: NumberFromString })
+})
+
+//
+// union
+//
+
+// $ExpectType TaskDecoder<string, string | number>
+_.union(NumberFromString, string)