+++ test/JsonEncoder.ts
@@ -2,9 +2,9 @@ import * as assert from 'assert'
 import * as JE from '../src/JsonEncoder'
 
 describe('Encoder', () => {
-  describe('contravariantJsonEncoder', () => {
+  describe('Contravariant', () => {
     it('contramap', () => {
-      const encoder = JE.contravariantJsonEncoder.contramap(JE.schemableJsonEncoder.number, (s: string) => s.length)
+      const encoder = JE.Contravariant.contramap(JE.Schemable.number, (s: string) => s.length)
       assert.deepStrictEqual(encoder.encode('aaa'), 3)
     })
   })