+++ test/Schema.ts
@@ -19,11 +19,11 @@ function isDeepStrictEqual(actual: unknown, expected: unknown): boolean {
 }
 
 function check<A>(schema: Schema<A>): void {
-  const arb = interpreter(A.schemableArbitrary)(schema)
-  const codec = interpreter(JC.schemableJsonCodec)(schema)
-  const guard = interpreter(G.schemableGuard)(schema)
-  const eq = interpreter(Eq.schemableEq)(schema)
-  const encoder = interpreter(JE.schemableJsonEncoder)(schema)
+  const arb = interpreter(A.Schemable)(schema)
+  const codec = interpreter(JC.Schemable)(schema)
+  const guard = interpreter(G.Schemable)(schema)
+  const eq = interpreter(Eq.Schemable)(schema)
+  const encoder = interpreter(JE.Schemable)(schema)
   // decoders, guards and eqs should be aligned
   fc.assert(fc.property(arb, (a) => isRight(codec.decode(a)) && guard.is(a) && eq.equals(a, a)))
   // laws